requirements.txt
fastapi
uvicorn
pandas
numpy
SQLAlchemy==2.0.17
PyMySQL==1.1.0
cloud-sql-python-connector==1.2.4
tqdm